+++ -d
@@ -1,0 +1,5 @@
+# Netscape HTTP Cookie File
+# https://curl.se/docs/http-cookies.html
+# This file was generated by libcurl! Edit at your own risk.
+
+#HttpOnly_localhost	FALSE	/	FALSE	1657536008	fEVR_Session	eyJfcGVybWFuZW50Ijp0cnVlfQ.Ysv3AA.nb-2fPLFOwCK9MyzB-wiSRfX5Eo
