+++ app/blueprints/api.py
@@ -24,11 +24,14 @@ from datetime import datetime
 import os
 import shutil
 
-from .models.models import events,frigate,cameras
-from . import db
-from .fetch import Fetch
-from .helpers.cookies import cookies
-from .helpers.iterateQuery import iterateQuery
+from app.models.frigate import frigate
+from app.models.events import events
+from app.models.frigate import frigate
+from app.models.cameras import cameras
+from app import db
+from app.helpers.fetch import Fetch
+from app.helpers.cookies import cookies
+from app.helpers.iterateQuery import iterateQuery
 
 # API Routes
 api = Blueprint('api',__name__)
