+++ app/mqtt_client
@@ -26,10 +26,19 @@ from random import randint
 import socket
 import requests
 from math import floor
-from logit import logit
 import socket
 import json
 import yaml
+import sys
+
+class logit:
+    def execute(msg,src='fEVR',debug=True):
+        def to_stderr(*a):
+            print(*a, file=sys.stderr)
+        logtime = "{:.4f}".format(time())
+        logentry = f"{logtime} {str(msg)}"
+        if debug:
+            to_stderr(f"[ {src:15}] {logentry}")
 
 class getLocalIP:
     s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -43,7 +52,6 @@ class auth:
         self.rSession = rSession
         self.apikey = apikey
         self.script = "MQTT | auth"
-        self.api()
 
     def api(self):
         # Construct the json for the post request
@@ -52,8 +60,9 @@ class auth:
         fevrUrl = f"{self.transport}{self.fevr}/apiAuth"
         apiAuth = self.rSession.post(f"{fevrUrl}", json=JsonPost)
         apiAuthJSON = json.loads(apiAuth.content)
-        authResult = "SUCCESS" if apiAuthJSON['auth'] else "FAILURE"
+        authResult = True if apiAuthJSON['auth'] else False
         logit.execute(f"{fevrUrl} AUTH {authResult}: {apiAuthJSON['authIP']} name: {apiAuthJSON['name']}",src=self.script)
+        return authResult
     
 class fevrMQTT:
     def __init__(self):
@@ -158,7 +167,7 @@ class fevrMQTT:
                 value = Key
             else:
                 value = self.config[key]
-            logit.execute(f"{key}: {value}")
+            logit.execute(f"{key}: {value}", src="MQTT | setup")
         # Parse the topics into a list if it's a string.
         if isinstance(self.config['mqtt_topics'],str):
             self.config['mqtt_topics'] = self.parseTopics()
@@ -193,14 +202,18 @@ class fevrMQTT:
             if event['type'] == 'end':
                 # If the event is over, we create a session variable, authenticate, and tell fEVR via it's API to get the event.
                 rSession = requests.Session()
-                auth(transport=self.config['fevr_transport'],apikey=self.config['mqtt_apikey'],fevr=self.config['fevr'],rSession=rSession)
-                fEVRurl = f"{self.config['fevr_transport']}{self.config['fevr']}/api/events/add/{event['after']['id']}/{event['after']['camera']}/{event['after']['label']}/{floor(float(event['after']['score'])*100)}"
-                addEventResponse = rSession.get(fEVRurl)
-                respJSON = json.loads(addEventResponse.content)
-                if self.config['verbose']:
-                    logit.execute(fEVRurl,src=self.script)
-                LogMsg = f"SUCCESS: {event['after']['id']} ({event['after']['label']} in {event['after']['camera']}) added" if respJSON["error"] == 0 else f"FAILURE: {respJSON}"
-                logit.execute(LogMsg,src=self.script)
+                Auth = auth(transport=self.config['fevr_transport'],apikey=self.config['mqtt_apikey'],fevr=self.config['fevr'],rSession=rSession)
+                if Auth.api():
+                    fEVRuri = f"{self.config['fevr_transport']}{self.config['fevr']}"
+                    fEVRurl = f"{fEVRuri}/api/events/add/{event['after']['id']}/{event['after']['camera']}/{event['after']['label']}/{floor(float(event['after']['score'])*100)}"
+                    addEventResponse = rSession.get(fEVRurl)
+                    respJSON = json.loads(addEventResponse.content)
+                    if self.config['verbose']:
+                        logit.execute(fEVRurl,src=self.script)
+                    LogMsg = f"SUCCESS: {event['after']['id']} ({event['after']['label']} in {event['after']['camera']}) added" if respJSON["error"] == 0 else f"FAILURE: {respJSON}"
+                    logit.execute(LogMsg,src=self.script)
+                else:
+                    logit.execute(f"Failed to Authenticate - visit {fEVRurl}/profile to generate an API Auth Key and insert it in your config", src="MQTT | ERROR")
                 
     def subscribe(self, client: mqtt_client):
         def on_message(client, userdata, msg):
