+++ app/static/css/main.css
@@ -443,6 +443,9 @@ a:hover{
     font-size: 1em;
     width: 98%;
 }
+.login input[type="checkbox" i] {
+    width: 1.25em;
+}
 .login .checkbox{
     height: 1em;
     width: 1em;
