--- docker-compose.full-stack.yml
@@ -1,100 +1,0 @@
-version: "2.4"
-services:
-  mqtt:
-    image: eclipse-mosquitto:latest
-    container_name: mqtt
-    restart: unless-stopped
-    ports:
-      - 1883:1883
-    volumes:
-      - ./mqtt/mosquitto.conf:/mosquitto/config/mosquitto.conf
-      - ./mqtt/data:/mosquitto/data
-      - ./mqtt/log:/mosquitto/log
-    networks:
-      fevr_bridge:
-          ipv4_address: ${MQTT_IP:-192.168.240.201}
-
-  rtsp-relay:
-    image: aler9/rtsp-simple-server
-    container_name: rtsp-relay
-    restart: unless-stopped
-    volumes:
-      - ./rtsp-relay/rtsp-simple-server.yml:/rtsp-simple-server.yml
-    ports:
-      - 5082:8554
-      - 5083:9997
-      - 5084:8888
-    environment:
-      RTSP_PROTOCOLS: tcp
-    networks:
-      fevr_bridge:
-          ipv4_address: ${RTSP_IP:-192.168.240.202}
-  frigate:
-    container_name: frigate
-    privileged: true # this may not be necessary for all setups
-    restart: unless-stopped
-    image: blakeblackshear/frigate:0.10.1-amd64
-    shm_size: "1024M" # update for your cameras based on calculation above
-    devices:
-      - /dev/bus/usb:/dev/bus/usb # passes the USB Coral, needs to be modified for other versions
-      - /dev/dri/renderD128 # for intel hwaccel, needs to be updated for your hardware
-    volumes:
-      - /etc/localtime:/etc/localtime:ro
-      - ./frigate/config.yml:/config/config.yml:ro
-      - /export/backup/frigate/media:/media/frigate
-    ports:
-      - "5000:5000"
-      - "1935:1935" # RTMP feeds
-    depends_on:
-      - mqtt
-      - rtsp-relay
-    environment:
-      FRIGATE_RTSP_PASSWORD: "X.27@nihilat0r"
-      #LIBVA_DRIVER_NAME: radeonsi
-      #NVIDIA_VISIBLE_DEVICES: all
-      #NVIDIA_DRIVER_CAPABILITIES: compute,utility,video
-    networks:
-      fevr_bridge:
-          ipv4_address: ${FRIGATE_IP:-192.168.240.200}
-  fevr:
-    image: docker.beardedtek.com/beardedtek/fevr:latest
-    container_name: fevr
-    restart: unless-stopped
-    privileged: true
-    ports:
-      - 5080:${FEVR_PORT:-5090}
-      - 5090:5090
-    volumes:
-      - /export/fevr:/fevr/app/static/events
-      - /export/fevr/data:/fevr/app/data
-      - ./fevr/varlib:/var/lib
-      - ./fevr/config:/fevr/app/data/config:ro
-    depends_on:
-      - mqtt
-      - rtsp-relay
-      - frigate
-    environment:
-      FEVR_DEVELOPMENT: ${FEVR_DEVELOPMENT:-false}
-      FEVR_URL: ${FEVR_URL}
-      FEVR_PORT: ${FEVR_PORT}
-      TAILSCALE_ENABLE: ${TAILSCALE_ENABLE:-true}
-      TAILSCALE_AUTHKEY: ${TAILSCALE_AUTHKEY}
-      TAILSCALE_HOSTNAME: ${TAILSCALE_HOSTNAME:-fevr}
-      TAILSCALE_TAGS: ${TAILSCALE_TAGS}
-      MQTT_BROKER: ${MQTT_BROKER:-mqtt}
-      MQTT_BROKER_PORT: ${MQTT_BROKER_PORT}
-      MQTT_BROKER_USER: ${MQTT_BROKER_USER}
-      MQTT_BROKER_PASSWORD: ${MQTT_BROKER_PASSWORD}
-      MQTT_TOPICS: ${MQTT_TOPICS:-frigate/+}
-      MQTT_VERBOSE_LOGGING: ${MQTT_VERBOSE_LOGGING:-true}
-      MQTT_APIAUTH_KEY: ${MQTT_APIAUTH_KEY}
-    networks:
-      fevr_bridge:
-          ipv4_address: ${TRANSMISSION_IP:-192.168.240.1}
-networks:
-  fevr_bridge:
-    driver: bridge
-    ipam:
-      config:
-        - subnet: ${DOCKER_SUBNET:-192.168.240.0/24}
-          gateway: ${DOCKER_GATEWAY:-192.168.240.254}
